big update
Added SettingsLoader.js which will store all settings in a single
cookie.Settings in general can be changed with :<setting> and printed
with 'printAddOnSettings. Added MirrorPlayer.js which will enable it to
mirror a video and also automaticly mirrors videos when the title
contains 'mirrored' or 'mirror' (this feature can be turned off with the
new settings), added nsfw emotes which can be enabled via settings. some
refactoring, improved autocomplete and inputhistory a bit